[feat]기능츄가

+++ pages/api/post/delete.js
@@ -1,13 +1,37 @@
 import { connectDB } from '@/util/database';
 import { ObjectId } from 'mongodb';
+import { authOptions } from '../auth/[...nextauth]';
+import { getServerSession } from 'next-auth';
 
 export default async function handler(요청, 응답) {
   if (요청.method == 'DELETE') {
     console.log(요청.body);
+    let session = await getServerSession(요청, 응답, authOptions);
+
     const db = (await connectDB).db('forum');
-    let result = await db
+    let 찾은거 = await db
       .collection('post')
-      .deleteOne({ _id: new ObjectId(요청.body) });
-    응답.status(200).json('삭제완료');
+      .findOne({ _id: new ObjectId(요청.body) });
+
+    if (찾은거.author == session.user.email) {
+      let result = await db
+        .collection('post')
+        .deleteOne({ _id: new ObjectId(요청.body) });
+      return 응답.status(200).json('삭제완료');
+    } else {
+      return 응답.status(500).json('현재유저와 작성자 불일치');
+    }
   }
 }
+// import { connectDB } from '@/util/database';
+// import { ObjectId } from 'mongodb';
+
+// export default async function handler(요청, 응답) {
+//   if (요청.method == 'DELETE') {
+//     const db = (await connectDB).db('forum');
+//     let result = await db
+//       .collection('post')
+//       .deleteOne({ _id: new ObjectId(요청.body) });
+//     응답.status(200).json('삭제완료');
+//   }
+// }
